+++ build.ps1
@@ -1,8 +1,10 @@
 $ErrorActionPreference = "Stop"
 
-dotnet test SentryNoSamples.slnf -c Release -l "GitHubActions;report-warnings=false" `
+$testLogger = if ($env:GITHUB_ACTIONS -eq "true") {"GitHubActions;report-warnings=false"} else {"console"}
+
+dotnet test SentryWindows.slnf -c Release -l $testLogger `
     /p:CopyLocalLockFileAssemblies=true
 if ($LASTEXITCODE -ne 0) { exit 1 }
 
-dotnet pack SentryNoSamples.slnf -c Release --no-build /p:ContinuousIntegrationBuild=true
+dotnet pack SentryWindows.slnf -c Release --no-build
 if ($LASTEXITCODE -ne 0) { exit 1 }